refactor(components): move profile button to user button wrapper
