+++ packages/nextjs/components/unlock/Utils.tsx
@@ -1,5 +1,5 @@
-import { ProfileButton } from "~~/components/Header/components/ProfileButton";
 import { RainbowKitCustomConnectButton } from "~~/components/Header/components/RainbowKitCustomConnectButton";
+import { ProfileButton } from "~~/components/UserButton/components/ProfileButton";
 import { useWeb3AuthContext } from "~~/context/Web3AuthContext";
 
 export const getTimestampOneMonthFromNow = (): number => {
